function that measures the height of a binary tree

+++ 9-binary_tree_height.c
@@ -11,8 +11,8 @@ size_t binary_tree_height(const binary_tree_t *tree)
 {
 	if (tree == NULL)
 		return (0);
-	int left_height = 0;
-	int right_height = 0;
+	size_t left_height = 0;
+	size_t right_height = 0;
 
 	if (tree->left)
 		left_height = binary_tree_height(tree->left) + 1;
